Move matchers to Shakedown.
Summary:
Let's move the matchers from the Marathon system tests to Shakedown so
that others can profit.

JIRA issues: MARATHON-8388
